Example life table: Austria, 1992, males

Example life table from Preston text book page 49. Post-hoc adjustment has been made to use radix (l0) of 1 instead of 100,000, as presented in the Preston book.

Original input data if from Austria, for males, from 1992, and comes from the United Nations, 1994, as cited in Preston book.

austria_1992_lt

Format

[data.table()] columns:

age_start

[integer()]
Abridged ages 0, 1, 5, 10, ..., 85. This is the 'x' variable in life table notation.

age_end

[integer()]
End of age group inclusive, 1, 5, 10, ... , 85, Inf

age_length

[integer()]
Width of the age interval in years. This is the 'n' variable in life table notation.

deaths

[numeric()]
Death counts in interval

population

[numeric()]
Population counts in interval

mx

[numeric()]
Mortality rate (deaths / population)

ax

[numeric()]
Mean person-years lived by those who die in interval.

qx

[numeric()]
Probability of death in interval conditional on survival to xth birthday.

px

[numeric()]
Probability of survival in interval.

lx

[numeric()]
Proportion surviving to age x.

dx

[numeric()]
Proportion of cohort dying in age interval x to x+n.

nLx

[numeric()]
Person-years lived between x and x+n per every one person in the cohort at birth.

Tx

[numeric()]
Person-years lived beyond age x per every one person in the cohort at birth.

ex

[numeric()]
Life expectancy.

Examples

austria_1992_lt

#>     age_start age_end age_length deaths    pop           mx    ax          qx
#>  1:         0       1          1    419  47925 0.0087428273 0.068 0.008672164
#>  2:         1       5          4     70 189127 0.0003701217 1.626 0.001479187
#>  3:         5      10          5     36 234793 0.0001533265 2.500 0.000766339
#>  4:        10      15          5     46 238790 0.0001926379 3.143 0.000962845
#>  5:        15      20          5    249 254996 0.0009764859 2.724 0.004871602
#>  6:        20      25          5    420 326831 0.0012850678 2.520 0.006404927
#>  7:        25      30          5    403 355086 0.0011349363 2.481 0.005658505
#>  8:        30      35          5    441 324222 0.0013601791 2.601 0.006778776
#>  9:        35      40          5    508 269963 0.0018817393 2.701 0.009368169
#> 10:        40      45          5    769 261971 0.0029354394 2.663 0.014577196
#> 11:        45      50          5   1154 238011 0.0048485154 2.698 0.023974985
#> 12:        50      55          5   1866 261612 0.0071327003 2.676 0.035081969
#> 13:        55      60          5   2043 181385 0.0112633349 2.645 0.054861466
#> 14:        60      65          5   3496 187962 0.0185995042 2.624 0.089061670
#> 15:        65      70          5   4366 153832 0.0283816111 2.619 0.132925406
#> 16:        70      75          5   4337 105169 0.0412383877 2.593 0.187573266
#> 17:        75      80          5   5279  73694 0.0716340543 2.518 0.304102201
#> 18:        80      85          5   6460  57512 0.1123243845 2.423 0.435548174
#> 19:        85     Inf        Inf   6146  32248 0.1905854627 5.247 1.000000000
#>            px      lx      dx     nLx       Tx        ex
#>  1: 0.9913278 1.00000 0.00867 0.99192 72.88902 72.889020
#>  2: 0.9985208 0.99133 0.00147 3.96183 71.89710 72.525900
#>  3: 0.9992337 0.98986 0.00076 4.94741 67.93527 68.631190
#>  4: 0.9990372 0.98910 0.00095 4.94375 62.98786 63.681994
#>  5: 0.9951284 0.98815 0.00481 4.92980 58.04411 58.740181
#>  6: 0.9935951 0.98334 0.00630 4.90106 53.11431 54.014186
#>  7: 0.9943415 0.97704 0.00553 4.87127 48.21325 49.346240
#>  8: 0.9932212 0.97151 0.00659 4.84175 43.34198 44.613004
#>  9: 0.9906318 0.96492 0.00904 4.80384 38.50023 39.899919
#> 10: 0.9854228 0.95588 0.01393 4.74686 33.69639 35.251695
#> 11: 0.9760250 0.94195 0.02258 4.65777 28.94953 30.733616
#> 12: 0.9649180 0.91937 0.03225 4.52188 24.29176 26.422180
#> 13: 0.9451385 0.88711 0.04867 4.32096 19.76988 22.285714
#> 14: 0.9109383 0.83845 0.07467 4.01480 15.44892 18.425571
#> 15: 0.8670746 0.76377 0.10152 3.57713 11.43412 14.970633
#> 16: 0.8124267 0.66225 0.12422 3.01224  7.85699 11.864085
#> 17: 0.6958978 0.53803 0.16362 2.28404  4.84475  9.004609
#> 18: 0.5644518 0.37441 0.16307 1.45182  2.56071  6.839321
#> 19: 0.0000000 0.21134 0.21134 1.10889  1.10889  5.246948