Estimate child mortality from sbh information

Estimate child mortality from sbh information from one source using the Brass method with Trussell coefficients.

calc_nqx_brass(
  dt,
  id_cols = c("age_start", "age_end"),
  parity_col = "Pi",
  prop_died_col = "Di",
  model_schedule = "West"
)

Arguments

dt

[data.table()]
sbh data with id_cols, parity_col, prop_died_col.

id_cols

[character()]
ID columns that uniquely identify each row of dt. This must include 'age_start' and 'age_end'.

parity_col

[character(1)]
Name of column storing the average parity for each maternal age group. Default is 'Pi'.

prop_died_col

[character(1)]
Name of column storing the proportion of children who have died for each maternal age group. Default is 'Di'.

model_schedule

[character(1)]
Name of model fertility schedule to use from sbh_trussell_coeffs. Default is "West". Other implemented option is "North".

Value

[data.table()] with id_cols, the estimated probability of dying ('nqx'), and the reference period ('tx').

Details

The method reproduced in this function is described in Section B of Chapter III in the UN Population Division Manual X. It estimates the probability of child mortality with data classified by maternal age from one survey or census.

Summary of steps:

Confirm average parity for each maternal age group ('P(i)') and proportion of children who have died for each maternal age group ('D(i)') are included in the input data.

Use estimation equations to calculate probability of child mortality: $$k(i)=a(i) + b(i) (P(1)/P(2)) + c(i) (P(2)/P(3))$$ $$q(x) = k(i)D(i)$$

Use estimation equation to calculate reference period: $$t(x)=a(i) + b(i) (P(1)/P(2)) + c(i) (P(2)/P(3))$$

References

UN Population Division. 1983. Manual X: Indirect Techniques for Demographic Estimation. New York: United Nations, Department of Economic and Social Affairs, ST/ESA/SER.A/81. http://www.un.org/esa/population/techcoop/DemEst/manual10/manual10.html

Examples

calc_nqx_brass(
  dt = sbh_panama_1976,
  id_cols = c("sex", "age_start", "age_end"),
  model = "West"
)
#> Key: <sex, age_start, age_end>
#>        sex age_start age_end        nqx         tx
#>     <char>     <num>   <num>      <num>      <num>
#>  1:    all         0       1 0.07658380  1.0480013
#>  2:    all         0       2 0.05137070  2.3648516
#>  3:    all         0       3 0.06517946  4.3149655
#>  4:    all         0       5 0.07178590  6.6365203
#>  5:    all         0      10 0.09670918  9.1944064
#>  6:    all         0      15 0.10881831 11.9241915
#>  7:    all         0      20 0.13087434 14.8558145
#>  8: female         0       1 0.05893272  1.1361667
#>  9: female         0       2 0.04403690  2.4070309
#> 10: female         0       3 0.05943267  4.2386973
#> 11: female         0       5 0.06758414  6.4061090
#> 12: female         0      10 0.09082286  8.7968066
#> 13: female         0      15 0.09647958 11.4040353
#> 14: female         0      20 0.13093046 14.3310511
#> 15:   male         0       1 0.09521042  0.9648135
#> 16:   male         0       2 0.05799854  2.3270195
#> 17:   male         0       3 0.07073958  4.3919097
#> 18:   male         0       5 0.07573630  6.8621158
#> 19:   male         0      10 0.10214683  9.5808416
#> 20:   male         0      15 0.12014862 12.4282225
#> 21:   male         0      20 0.13078712 15.3637414
#>        sex age_start age_end        nqx         tx
calc_nqx_brass(
  dt = sbh_panama_1976,
  id_cols = c("sex", "age_start", "age_end"),
  model = "North"
)
#> Key: <sex, age_start, age_end>
#>        sex age_start age_end        nqx         tx
#>     <char>     <num>   <num>      <num>      <num>
#>  1:    all         0       1 0.07476717  1.0420828
#>  2:    all         0       2 0.04905409  2.3076415
#>  3:    all         0       3 0.06206615  4.1421502
#>  4:    all         0       5 0.07015629  6.3293016
#>  5:    all         0      10 0.09872205  8.7764823
#>  6:    all         0      15 0.11096757 11.4280222
#>  7:    all         0      20 0.13205485 14.3135882
#>  8: female         0       1 0.05727491  1.1289033
#>  9: female         0       2 0.04200717  2.3498209
#> 10: female         0       3 0.05672484  4.0770878
#> 11: female         0       5 0.06629113  6.1247792
#> 12: female         0      10 0.09303494  8.4130733
#> 13: female         0      15 0.09861116 10.9358628
#> 14: female         0      20 0.13217995 13.7925362
#> 15:   male         0       1 0.09332109  0.9601607
#> 16:   male         0       2 0.05543359  2.2696791
#> 17:   male         0       3 0.06720329  4.2080930
#> 18:   male         0       5 0.07374867  6.5297947
#> 19:   male         0      10 0.10391542  9.1299190
#> 20:   male         0      15 0.12224438 11.9051178
#> 21:   male         0      20 0.13190021 14.8180478
#>        sex age_start age_end        nqx         tx