Make Leslie matrix

Constructs the Leslie matrix needed for cohort component method of population projection ccmpp().

leslie_matrix(survival, asfr, srb, n_ages, int, female)

Arguments

survival

[numeric(n_ages + 1) or numeric(n_ages)]
Survivorship ratio, the proportion of people aged x - int that will be alive int years later in a stationary population.

asfr

[numeric(n_ages)]
Annual age specific fertility rates NOT yet multiplied by int. Must include both reproductive and non-reproductive age groups that are zero.

srb

[numeric(1)]
Sex ratio at birth.

n_ages

[integer(1)]
Number of age groups.

int

[integer(1)]
Width of the age groups and projection intervals. Usually 1 or 5.

female

[logical(1)]
Whether making Leslie matrix for projection of the female population (versus male population).

Value

[matrix(n_ages, n_ages)]

Details

The Leslie matrix encodes two key pieces of information:

The survivorship ratio is included in the off diagonal to age the population in each age group \(a\) at time \(t\) to age group \(a + int\) at time \(t + int\).

When projecting the female population, the first row includes information about age-specific fertility rates, maternal survivorship ratios, birth survivorship ratios and the sex-ratio at birth to calculate the initial population size of the youngest age group at time \(t + int\).

See the references for more information.

References

Preston, Samuel, Patrick Heuveline, and Michel Guillot. 2001. Demography: Measuring and Modeling Population. Wiley.

Examples

leslie <- leslie_matrix(
  survival = thailand_initial_estimates$survival[year_start == 1960 &
                                                  sex == "female", value],
  asfr = c(rep(0, 3),
           thailand_initial_estimates$asfr[year_start == 1960, value],
           rep(0, 7)),
  srb = thailand_initial_estimates$srb[year_start == 1960, value],
  n_ages = 17, int = 5, female = TRUE
)