Thailand example population data and ccmpp initial estimates

Example data and initial estimates for reconstructing the female and male population in Thailand from 1960 to 2000.

thailand_data

thailand_initial_estimates

Format

thailand_data: [data.table()] with 'population' data.

population: [data.table()] year-sex-age-specific census counts in years after the baseline year (1970, 1980, 1990, 2000). Age groups are five-year age groups from 0 to 80+.

thailand_initial_estimates: list of [data.table()] of initial estimates for each ccmpp() input. Calendar year intervals are for five-year intervals between 1960 and 2000. Age groups are five-year age groups from 0 to 80+ (except 'survival' goes up to 85+). See Section: Inputs for more information on each of the inputs.

An object of class list of length 5.

Inputs

srb: [data.table()]

year_start: [integer()] start of the calendar year interval (inclusive). Corresponds to 'years' setting.

year_end: [integer()] end of the calendar year interval (exclusive).

value_col: [numeric()] sex-ratio at birth estimates, must be greater than zero.

asfr: [data.table()]

year_start: [integer()] start of the calendar year interval (inclusive). Corresponds to 'years' setting.

year_end: [integer()] end of the calendar year interval (exclusive).

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages_asfr' setting.

age_end: [integer()] end of the age group (exclusive).

value_col: [numeric()] annual age-specific fertility rate estimates, must be greater than zero.

baseline: [data.table()]

year: [integer()] mid-year for population estimate. Corresponds to 'years' setting.

sex: [character()] either 'female' or 'male'. Corresponds to 'sexes' setting.

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages' setting.

age_end: [integer()] end of the age group (exclusive).

value_col: [numeric()] baseline year population count estimates, must be greater than zero.

survival: [data.table()]

year_start: [integer()] start of the calendar year interval (inclusive). Corresponds to 'years' setting.

year_end: [integer()] end of the calendar year interval (exclusive).

sex: [character()] either 'female' or 'male'. Corresponds to 'sexes' setting.

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages' setting.

age_end: [integer()] end of the age group (exclusive).

value_col: [numeric()] survivorship ratio estimates, must be greater than zero and less than one.

net_migration: [data.table()]

year_start: [integer()] start of the calendar year interval (inclusive). Corresponds to 'years' setting.

year_end: [integer()] end of the calendar year interval (exclusive).

sex: [character()] either 'female' or 'male'. Corresponds to 'sexes' setting.

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages' setting.

age_end: [integer()] end of the age group (exclusive).

value_col: [numeric()] annual net-migration proportion estimates.

immigration: [data.table()]

year_start: [integer()] start of the calendar year interval (inclusive). Corresponds to 'years' setting.

year_end: [integer()] end of the calendar year interval (exclusive).

sex: [character()] either 'female' or 'male'. Corresponds to 'sexes' setting.

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages' setting.

age_end: [integer()] end of the age group (exclusive).

value_col: [numeric()] annual immigration proportion estimates, must be greater than zero.

emigration: [data.table()]

year_start: [integer()] start of the calendar year interval (inclusive). Corresponds to 'years' setting.

year_end: [integer()] end of the calendar year interval (exclusive).

sex: [character()] either 'female' or 'male'. Corresponds to 'sexes' setting.

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages' setting.

age_end: [integer()] end of the age group (exclusive).

value_col: [numeric()] annual emigration proportion estimates, must be greater than zero.

References

Wheldon, Mark C., Adrian E. Raftery, Samuel J. Clark, and Patrick Gerland. 2015. “Bayesian Reconstruction of Two-Sex Populations by Age: Estimating Sex Ratios at Birth and Sex Ratios of Mortality.” Journal of the Royal Statistical Society. Series A: Statistics in Society 178 (4): 977–1007. https://doi.org/10.1111/rssa.12104.

markalava/Bayesian-Reconstruction github repo

See also