Validate the ccmpp input arguments

Validate the ccmpp input arguments

validate_ccmpp_inputs(inputs, settings, value_col)

Arguments

inputs

[list()]
[data.table()] for each ccmpp input. Requires 'srb', 'asfr', 'baseline'; mortality estimates provided as just 'survival' or two of 'mx', 'ax', and 'qx'; and migration estimates provided as just 'net_migration' or both 'immigration' and 'emigration'. See Section: Inputs for more information on each of the required inputs.

settings

[list()]
Named list of settings for running ccmpp() with. See Section: Settings for more information on each of the required settings.

value_col

[character(1)]
Name of the column containing the value of interest in each of the inputs. Default is 'value'.

Value

Invisibly returns inputs but throws error if not formatted correctly.