Regression parameters for abridged to full life table conversion

Parameters for age- and sex-specific regression model fit to Human Mortality Database data, to predict single-year qx values from abridged qx values: $$log(q_x) = intercept + slope*log(q_x,abridged)$$

full_lt_pars

Format

[data.table()] columns:

sex

[character()]
Sex the parameters correspond to: "male" or "female"

age_start

[integer()]
Single-year ages 1-109 representing the age the parameters correspond to

age_end

[integer()]
1 + age_start

intercept

[numeric()]
Intercept for the age- and sex-specific regression model fit

slope

[numeric()]
Slope for the age- and sex-specific regression model fit

References

Human Mortality Database. University of California, Berkeley (USA), and Max Planck Institute for Demographic Research (Germany). Available at www.mortality.org or www.humanmortality.de (data downloaded on Jan 2018).

Examples

full_lt_pars

#>         sex age_start age_end  intercept     slope
#>   1:   male         1       2 -0.5176810 1.0584153
#>   2: female         1       2 -0.5909254 1.0341440
#>   3:   male         2       3 -1.5681676 0.9753858
#>   4: female         2       3 -1.5817470 0.9749719
#>   5:   male         3       4 -2.1365209 0.9249339
#>  ---                                              
#> 214: female       107     108 -0.6030596 2.5503541
#> 215:   male       108     109 -0.5668620 2.6699057
#> 216: female       108     109 -0.5757869 2.5567061
#> 217:   male       109     110 -0.5452189 2.6738654
#> 218: female       109     110 -0.5510617 2.5584597